refactoring, renamed the Notes class to NotesImpl and made the Notes interface
