+++ app/src/main/java/ru/androidlearning/notes/ui/NoteTitlesAdapter.java
@@ -11,6 +11,7 @@ import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.RecyclerView;
 
 import ru.androidlearning.notes.R;
+import ru.androidlearning.notes.data.NoteEntry;
 import ru.androidlearning.notes.data.Notes;
 
 public class NoteTitlesAdapter extends RecyclerView.Adapter<NoteTitlesAdapter.ViewHolder> {
@@ -90,13 +91,10 @@ public class NoteTitlesAdapter extends RecyclerView.Adapter<NoteTitlesAdapter.Vi
             }
         }
 
-        public void setData(Notes.NoteEntry noteEntry) {
+        public void setData(NoteEntry noteEntry) {
             noteTitle.setText(noteEntry.getNoteTitle());
             noteDate.setText(noteEntry.getNoteCreatedDate());
             noteDetails.setText(noteEntry.getNoteText());
         }
-
-
     }
-
 }
